Move User interface to shared
