+++ src/main/java/uk/gov/mca/beacons/api/auth/application/GetUserService.java
@@ -3,7 +3,7 @@ package uk.gov.mca.beacons.api.auth.application;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import uk.gov.mca.beacons.api.auth.gateway.AuthGateway;
-import uk.gov.mca.beacons.api.domain.User;
+import uk.gov.mca.beacons.api.shared.domain.user.User;
 
 @Service
 public class GetUserService {
