+++ src/main/java/uk/gov/mca/beacons/api/auth/gateway/AuthGateway.java
@@ -2,7 +2,7 @@ package uk.gov.mca.beacons.api.auth.gateway;
 
 import java.util.List;
 import uk.gov.mca.beacons.api.auth.gateway.AuthGatewayImpl.SupportedPermissions;
-import uk.gov.mca.beacons.api.domain.User;
+import uk.gov.mca.beacons.api.shared.domain.user.User;
 
 public interface AuthGateway {
   User getUser();
