+++ src/main/java/uk/gov/mca/beacons/api/controllers/NoteController.java
@@ -1,17 +1,18 @@
 package uk.gov.mca.beacons.api.controllers;
 
 import io.swagger.v3.oas.annotations.tags.Tag;
+import java.util.UUID;
 import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import uk.gov.mca.beacons.api.auth.application.GetUserService;
 import uk.gov.mca.beacons.api.domain.Note;
-import uk.gov.mca.beacons.api.domain.User;
 import uk.gov.mca.beacons.api.dto.NoteDTO;
 import uk.gov.mca.beacons.api.dto.WrapperDTO;
 import uk.gov.mca.beacons.api.mappers.NoteMapper;
 import uk.gov.mca.beacons.api.services.NoteService;
+import uk.gov.mca.beacons.api.shared.domain.user.User;
 
 @RestController
 @RequestMapping("/spring-api/note")
@@ -41,7 +42,7 @@ public class NoteController {
     final Note note = noteMapper.fromDTO(dto.getData());
     final User user = getUserService.getUser();
 
-    note.setUserId(user.getId());
+    note.setUserId(UUID.randomUUID());
     note.setFullName(user.getFullName());
     note.setEmail(user.getEmail());
