--- src/main/java/uk/gov/mca/beacons/api/domain/User.java
@@ -1,9 +1,0 @@
-package uk.gov.mca.beacons.api.domain;
-
-import java.util.UUID;
-
-public interface User {
-  UUID getId();
-  String getFullName();
-  String getEmail();
-}
