+++ src/main/java/uk/gov/mca/beacons/api/shared/domain/user/User.java
@@ -1,0 +1,6 @@
+package uk.gov.mca.beacons.api.shared.domain.user;
+
+public interface User {
+  String getFullName();
+  String getEmail();
+}
